Add variable 'next_task'
An additional variable 'next_task' is helpful and clear for those who
wants to change something according to the next task.

If only 'current_task' is used, it must implicitly assume that the next
task is the current task before kernel selects the next one.

It is a fault when more than 2 tasks are created, for example:

　if Task [length of execution] => 1[1T], 2[2T], 3[3T]
　Task :　　　　　　　1 -> 2 -> 3 -> 1 -> 2 ->3
　Length of execution : 1T->2T->1T->2T->3T->1T (<--- which is wrong)